% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyecto_final.R
\name{crear_proyecto_final}
\alias{crear_proyecto_final}
\title{Crear proyecto final del Taller II}
\usage{
crear_proyecto_final(path, open = TRUE, with_git = FALSE)
}
\arguments{
\item{path}{Nombre del directorio en el que se creará el proyecto.
Será usado como nombre del proyecto.}

\item{open}{Boolean. Abrir el proyecto creado?}

\item{with_git}{Boolean. Iniciar proyecto con repositorio git?}
}
\value{
path
}
\description{
Crear proyecto final del Taller II
}
